rename inferencer
