+++ mmaction/apis/inferencers/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .actionrecog_inferencer import ActionRecogInferencer
-from .mmaction_inferencer import MMAction2Inferencer
+from .mmaction2_inferencer import MMAction2Inferencer
 
 __all__ = ['ActionRecogInferencer', 'MMAction2Inferencer']
